+++ packages/overlay/src/types.ts
@@ -112,5 +112,16 @@ export type WindowWithSpotlight = Window & {
 };
 
 export type SidePanelProps = Omit<ComponentPropsWithoutRef<'div'>, 'className'> & {
-  backTo: string;
+  backto: string;
+};
+
+export type CrumbProps = {
+  id: string;
+  label: string;
+  link?: boolean;
+  to?: string;
+};
+
+export type BreadcrumbProps = {
+  crumbs: CrumbProps[];
 };